+++ Engine/Core/SBOMGenerator.h
@@ -104,11 +104,11 @@ class SBOMGeneratorEngine
            << "  \"spdxVersion\": \"SPDX-2.3\",\n"
            << "  \"dataLicense\": \"CC0-1.0\",\n"
            << "  \"SPDXID\": \"SPDXRef-DOCUMENT\",\n"
-           << "  \"name\": \"ExplorerLens-32.6.0\",\n"
+           << "  \"name\": \"ExplorerLens-32.6.1\",\n"
            << "  \"documentNamespace\": \"https://explorerlensteam.github.io/spdx/" << uuid << "\",\n"
            << "  \"creationInfo\": {\n"
            << "    \"created\": \"" << ts << "\",\n"
-           << "    \"creators\": [\"Tool: ExplorerLens-SBOMGenerator-32.6.0\"]\n"
+           << "    \"creators\": [\"Tool: ExplorerLens-SBOMGenerator-32.6.1\"]\n"
            << "  },\n"
            << "  \"packages\": [\n";